backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
aiofiles==23.2.1
pymupdf==1.25.1
google-generativeai==0.3.2
deep-translator==1.11.4
arabic-reshaper==3.0.0
python-bidi==0.4.2
setuptools==69.0.2
dataclasses-json==0.6.3
transformers==4.35.2
torch==2.1.0 --index-url https://download.pytorch.org/whl/cpu
sentencepiece==0.1.99
websockets==12.0
requests==2.31.0
regex==2023.10.3
sacremoses==0.0.53
